+++ src/Contract/UserTokenInterface.php
@@ -3,7 +3,7 @@
 namespace Heptacom\AdminOpenAuth\Contract;
 
 use Heptacom\AdminOpenAuth\Database\UserTokenEntity;
-use Heptacom\AdminOpenAuth\Struct\TokenPairStruct;
+use Heptacom\OpenAuth\Struct\TokenPairStruct;
 use Shopware\Core\Framework\Context;
 
 interface UserTokenInterface